Change Card panel to be a carousel.
