+++ app/components/Header/index.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 const Header = styled.header`
   width: 100%;
+  flex-basis: 100%;
 `;
 
 export default Header;
