+++ app/global-styles.js
@@ -68,7 +68,7 @@ injectGlobal`
     --pin-size: 30px;
 
     @media (min-width: 720px) {
-      --topbar-height: 80px;
+      --topbar-height: 60px;
     }
   }
